MS bug with new libpng-1.2.6 files
There's something weird with the windowBits.  I suspect "8K" (i.e.
13 - 5 in CINFO) is wrong, because when I call inflateInit2 with that
value (i.e. 13) zlib 1.1.4 from pngcheck gives me:

zlib:  inflate (first loop) error = -3

That's "Z_DATA_ERROR".

There's a patch for pngcheck 1.99.3 at the end of this email to make it
test the value of CINFO (by reading it from the zlib 2 byte header and
feeding it back into inflateInit2).

The salient thing here is that both Office and Windows are using CINFO
to reduce the memory overhead of inflate (by causing it to allocate a
reduced window.)