fix signedness comparison issue
Before the change, one gets the following warning on GCC 11:

	cc  -c  -I../zlib -DUSE_ZLIB -pedantic-errors  -Wall -Wextra -Wundef  -O2  -o pngcheck.o pngcheck.c
	pngcheck.c: In function ‘pngcheck’:
	pngcheck.c:1964:36: warning: comparison of integer expressions of different signedness: ‘unsigned int’ and ‘int’ [-Wsign-compare]
	 1964 |                 if (i == Z_OK || i == Z_STREAM_ERROR)
		  |                                    ^~

Z_STREAM_ERROR is defined to be (-2), hence the warning.
cf. https://www.zlib.net/manual.html

Because of this v4.0.0 fails to compile out of the box on systems using recent
GHCs as the former Makefile.unx has `-Werror` active.

This change uses a different variable to store the result of `inflateEnd`.

Since v3.0.3, the variable `i` in main has been changed from signed to
unsigned.  However, inflateEnd returns an `int` not an `unsigned int`, so
reusing the `i` variable as the result doesn't seem proper.