Require the zlib library as a non-optional dependency
Since the inception of this project, the use of zlib as a dependency
has been optional. Both zlib and the broader PNG project itself were
also in their early stages of development.

Over time, the set of differences between pngcheck-with-zlib and
pngcheck-without-zlib has steadily increased, adding complexity to
development, testing and maintenance.

We believe it's time to unify pngcheck into a single, fully-functional
program in which the zlib library (or a compatible alternative) is a
strict requirement.