Update the wildargs subproject, fixing a missed compiler+runtime combo
Clang on Windows using the Visual C++ Runtime was already covered by
the previous version of wildargs; and Clang on MinGW, too, as long as
it didn't pretend that it was Microsoft C.

Some of the Clang and Clang-based compilers pretend to be GCC, while
others (Clang-CL in particular) pretend to be Microsoft C. This might
be useful to some projects. On the other hand, unfortunately for the
code logic used inside wildargs, this is actually detrimental: it can
mislead the detection of the accompanying runtime.

The wildargs project was fixed upstream, resolving the aforementioned
runtime detection issue by rewriting and reordering the preprocessor
checks. The README file was refined, too.

In this commit, we sync it up.

Signed-off-by: Cosmin Truta <ctruta@gmail.com>